refactor: alterando tipo do atributo monthlyPayment de Integer para Double nos controllers

+++ src/main/java/com/greenbeansapps/myschooltransportation/main/controllers/CreateStudentController.java
@@ -83,7 +83,7 @@ public class CreateStudentController {
           @NotBlank String school,
           @NotBlank String grade,
           @NotBlank String transportationType,
-          @NotNull Integer monthlyPayment,
+          @NotNull Double monthlyPayment,
           @NotNull Integer monthlyPaymentExpiration,
           @NotBlank String shift,
           @ValidUUID
